+++ src/components/Event/docs/index.story.tsx
@@ -53,6 +53,9 @@ export default {
           ...[{ title: 'More Events', source: examples.moreEventsSection }].map(
             code,
           ),
+          ...[{ title: 'Clickable', source: examples.clickableSection }].map(
+            code,
+          ),
         ],
       }),